+++ src/module.js
@@ -14,7 +14,7 @@
 import { registerSettings } from "./scripts/settings.js";
 import { initHooks, readyHooks, setupHooks } from "./scripts/raritycolors.js";
 import CONSTANTS from "./scripts/constants.js";
-import { error } from "./scripts/lib/lib.js";
+import Logger from "./scripts/lib/Logger.js";
 
 /* ------------------------------------ */
 /* Initialize module					*/
@@ -27,6 +27,11 @@ Hooks.once("init", () => {
   //         word = "activate";
   //     throw Logger.error(`Requires the 'libWrapper' module. Please ${word} it.`);
   // }
+  if (!game.modules.get("colorsettings")?.active && game.user?.isGM) {
+    let word = "install and activate";
+    if (game.modules.get("colorsettings")) word = "activate";
+    throw Logger.error(`Requires the 'colorsettings' module. Please ${word} it.`);
+  }
   // Register custom module settings
   registerSettings();